ImportCommand method for embedded images
