+++ src/commands/item/import.js
@@ -249,54 +249,18 @@ export class Import extends ImportCommand {
 
     let { list: activeList } = this.action.payload
 
-    info(`extracting ${embedded.length} embedded file(s) ` +
-      `from pdf portfolio "${source.path}"`)
-
-    let images = []
-    for (let { name, data, mimetype } of embedded) {
-      try {
-        let img = yield call(Image.fromBuffer, {
-          buffer: data,
-          mimetype,
-          filename: name,
-          source,
-          useLocalTimezone
-        })
-        images.push(img)
-      } catch (err) {
-        warn({ err, name }, 'skipping unreadable embedded pdf file')
-      }
-    }
+    let { images, imageData } = yield * this.prepareEmbeddedImages(
+      source,
+      embedded, {
+        store,
+        useLocalTimezone,
+        optimizeOnImport,
+        optimizeQuality
+      })
 
     if (!images.length) return
 
     let data = yield * this.getMetadata(images[0], templates)
-    let imageData = new Array(images.length)
-
-    for (let i = 0; i < images.length; i++) {
-      let image = images[i]
-
-      try {
-        if (optimizeOnImport && image.mimetype !== MIME.JPG) {
-          yield call([image, image.optimize], { quality: optimizeQuality })
-        }
-        yield * this.handleDuplicate(image)
-        yield call(store.add, image)
-        imageData[i] = image.toJSON()
-
-      } catch (err) {
-        if (err instanceof DuplicateError) {
-          info(`skipping duplicate "${image.filename}" in "${source.path}"`)
-          imageData[i] = null
-        } else {
-          throw err
-        }
-      }
-    }
-
-    if (imageData.every(d => d == null))
-      throw new DuplicateError(source.path)
-
     let item
     let photos = []
